Move Github tools to common
