+++ src/coder/lg_server.py
@@ -2,7 +2,7 @@
 
 from coder.graph import graph_builder
 from coder.mocks import MockGithubApi
-from coder.tools import get_github_tools
+from common.components.github_tools import get_github_tools
 
 # graph for LangGraph Server
 # TODO: maybe we can setup this to use mock/real github api depending on env/config
